Interact with pager subprocess directly
Rather than needing to use convoluted multiprocess stuff; this was more
complicated and the `less` processes were not being killed correctly, as
they do not seem to die when the parent process is killed.

Had to copy and adapt some Click code to do this but this seems simpler
overall.